Update: move prmtop tests to core and structure_io
