+++ enzy_htp/_interface/amber_interface.py
@@ -2187,82 +2187,6 @@ class AmberInterface(BaseInterface):
 
         return result
 
-    def parse_fmt(self, fmt: str) -> Tuple[Any, int]:
-        """Private helper function that parses a %FORMAT() string from an amber .prmtop file.
-        Identifies both the type and width of the format. Supplied string can contain whitespace.
-        In the case that the supplied format string does not contain the pattern '%FORMAT(<>)',
-        the result (None, -1) is returned.
-
-        Args:
-            fmt: the %FORMAT() line from a .prmtop as a str().
-
-        Returns:
-            A tuple() of length 2 containing the (type ctor, width as an int()).
-        """
-
-        if fmt.find('%FORMAT(') == -1 or fmt.find(')') == -1:
-            return (None, -1)
-
-        fmt = fmt.strip().replace('%FORMAT(', '').replace(')', '')
-
-        idx = 0
-
-        while fmt[idx].isnumeric():
-            idx += 1
-
-        dtype = None
-
-        type_string = fmt[idx].upper()
-
-        if type_string == 'A':
-            dtype = str
-        elif type_string == 'I':
-            dtype = int
-        elif type_string == 'E':
-            dtype = float
-
-        width = int(fmt[idx + 1:].split('.')[0])
-
-        return (dtype, width)
-
-    def parse_prmtop(self, prmtop: str) -> Dict:
-        """Parses a .prmtop file into a dict() with the pattern of (key, value) being (keyword after %FLAG, following values). 
-        Takes the specified %FORMAT() into account. Checks if the supplied file exists, exiting if it does not.
-
-        Args:
-            prmtop: str() holding the path to a .prmtop file.
-
-        Returns:
-            A dict() with (key, value) pairs of (keyword, formatted/typed values).
-
-        """
-        #TODO(CJ): unit tests
-
-        if not Path(prmtop).exists():
-            _LOGGER.error(f"The file '{prmtop}' does not exist. Exiting...")
-            exit(1)
-
-        raw: str = fs.content_from_file(prmtop)
-
-        result = dict()
-
-        for chunk in raw.split('%FLAG'):
-            if chunk.find('%VERSION') != -1:
-                #TODO(CJ): figure out what I want to do with version info
-                continue
-
-            (key, fmt, body) = chunk.split('\n', 2)
-            body = re.sub('\n', '', body)
-            (dtype, width) = self.parse_fmt(fmt)
-
-            raw: List[str] = list()
-            for idx in range(0, len(body), width):
-                raw.append(body[idx:idx + width].strip())
-
-            result[key.strip()] = list(map(dtype, raw))
-
-        return result
-
     def add_charges(self, stru: Structure, prmtop: str) -> None:
         """Method that adds RESP charges from a .prmtop file to the supplied Structure object. If the supplied prmtop
         file does not line up with the structure, an error is thrown. Performs operation in place.
